Created new spawner script to spawn fruits
Script takes a Fruit prefab as well as the minimum spawn time, maximum spawn time, minimum force and maximum force to impulse the fruit out of the spawner objects.

The script also contains a public property for a list of transform objects used to pass in the positions of the spawn places. (Fruits will spawned from the passed spawn places)

This script is assigned to empty spawner objects in the main scene.