Added multiple prefabs
Added prefabs for the following.

Watermelon, Cut Watermelon
Banana, Cut Banana
Bomb

All the left and right pieces of the cut prefabs contain a rigid body as well as a box collider.

All regular pieces of fruit and the bomb have a RigidBody2d component as well as a Circle collider 2d component. The Is Trigger property in the Circle Colliders is checked (Set to true)