Modified the spawner script
Changed the fruitToSpawn property into an array of objects to spawn.

Changed the line of code that instantiates the objects to create a random object using the array's index to select it.

Will change this logic into using a 10% probability to spawn a bomb. (Might tweak this value later or even make it public for easy changing).