Various changes to the Spawner Script
Changed the objectsToSpawn array into a fruitsToSpawn array. The fruits will be added in this array.

Added a new public Game Object to add the bomb.

Did this to avoid confusion if there came a time when someone else wanted to work with this code.

Modified the spawner script to have a 20% probability of spawning a bomb.

If the probability variable is 20 or below, the gameObj to spawn is a bomb.

Else it will spawn a random fruit.