Multiple Changes to the GameManager script
Created a method to restart the game.
I originally intended to load the scene again but used another method which involves tagging the Interactable objects and destroying them with a foreach loop as well as resetting the values such as score, scoreText, etc.

Imported the Scene Management package (might remove since another way to restart the game was implemented)

Added Headers to the GameManager properties. These headers will show in the Inspector when adding game objects to the scripts.

Added code to hide and show the game over panel whenever the game is over or is restarted.