Multiple Updates to GameManager Script
Added 4 new properties.

- A public Array of Audio Clips for the slicing sounds.
- A public audio clip for the Object spawn
- A public audio clip for the bomb explosion
- A private AudioSource to play all the audio clips with the PLayOneStop() method.

Added code to add advertisements.
Imported the UnityEngine.Advertisements package.
However, this code is commented and will not be a part of the project in GitHub. (If you wish to experiment with adding ads, create your own unity account and go through the process of getting your personal GameId for the project.)

Added methods to play the sounds from the passed audio clips. These include the spawn sound, slice sounds and explosion sound. Might add some background music later.

(Important)
Modified the OnBombHit method to show the GameOverPanel when the bomb explosion is heard.

Had to research how to do it. Tried to use the Invoke method but found out that Unity recommends using a coroutine for this sort of stuff.

Found out that the yield return new WaitForSeconds() wont work if the timescale is set to 0 so I set it to 0.01, multiplied that with 1.3 (which is the seconds it takes for the explosion to sound after the bomb-omb sound) and got 0.013 which is what I passed to the WaitForSeconds method. Proud of myself for this one. Perseverance wins.